kv-scanner: move to lib
For general availability kv-scanner moved from modules to lib/scanner
next to other scanners.
