+++ lib/scanner/kv-scanner/CMakeLists.txt
@@ -1,0 +1,7 @@
+set(KV_SCANNER_HEADERS
+    kv-scanner/kv-scanner.h
+    PARENT_SCOPE)
+
+set(KV_SCANNER_SOURCES
+    kv-scanner/kv-scanner.c
+    PARENT_SCOPE)
