+++ lib/scanner/kv-scanner/Makefile.am
@@ -1,0 +1,9 @@
+kvscannerincludedir			= ${pkgincludedir}/scanner/kv-scanner
+
+kvscannerinclude_HEADERS = 			\
+	lib/scanner/kv-scanner/kv-scanner.h
+
+kvscanner_sources = 				\
+	lib/scanner/kv-scanner/kv-scanner.c
+
+include lib/scanner/kv-scanner/tests/Makefile.am
