+++ lib/scanner/kv-scanner/kv-scanner.c
