+++ lib/scanner/kv-scanner/tests/test_kv_scanner.c
