+++ modules/kvformat/Makefile.am
@@ -5,8 +5,6 @@ modules_kvformat_libkvformat_la_SOURCES=	\
 	modules/kvformat/format-welf.h		\
 	modules/kvformat/kv-parser.c		\
 	modules/kvformat/kv-parser.h		\
-	modules/kvformat/kv-scanner.c		\
-	modules/kvformat/kv-scanner.h		\
 	modules/kvformat/linux-audit-parser.c	\
 	modules/kvformat/linux-audit-parser.h	\
 	modules/kvformat/kv-parser-grammar.y	\
