+++ modules/kvformat/tests/Makefile.am
@@ -3,15 +3,6 @@ modules_kvformat_tests_TESTS		= \
 	modules/kvformat/tests/test_linux_audit_scanner	\
 	modules/kvformat/tests/test_kv_parser
 
-if ENABLE_CRITERION
-modules_kvformat_tests_TESTS	+= modules/kvformat/tests/test_kv_scanner
-modules_kvformat_tests_test_kv_scanner_CFLAGS	= $(TEST_CFLAGS) -I$(top_srcdir)/modules/kvformat
-modules_kvformat_tests_test_kv_scanner_LDADD	= $(TEST_LDADD)
-modules_kvformat_tests_test_kv_scanner_LDFLAGS	= \
-	-dlpreopen $(top_builddir)/modules/kvformat/libkvformat.la
-modules_kvformat_tests_test_kv_scanner_DEPENDENCIES = $(top_builddir)/modules/kvformat/libkvformat.la
-endif
-
 check_PROGRAMS				+= ${modules_kvformat_tests_TESTS}
 
 modules_kvformat_tests_test_format_welf_CFLAGS	= $(TEST_CFLAGS)
